==> notebooks/01_data_loading_and_overview.ipynb <==
# Overview
- Load statistical metrics from EDF files and transform into a dataframe
- Observe dataframe dimension and data types
- Save dataframe to csv file

# Import the necessary packages from utils file
from utils import pd, np, mne, os

# Load one EDF file (PhysioNet uses .edf format)
path_name = '../data/raw/'

# Get all EDF files from the raw directory
edf_files = [f for f in os.listdir(path_name) if f.endswith('.edf')]

# Create a table of features
features = []
labels = []

# Create column names for features
feature_names = ['trial_id','subject_id']
for ch in range(1, 65):  # 64 channels
    feature_names.extend([
        f'ch{ch}_std',
        f'ch{ch}_mean',
        f'ch{ch}_max',
        f'ch{ch}_min'
    ])

# Process each file
raw_data = []
for file_name in edf_files:
    raw = mne.io.read_raw_edf(path_name + file_name, preload=True, encoding='latin1')
    raw.filter(1., 40.)  # Bandpass filtering
    raw_data.append(raw)

    # Assume annotations are already present
    events, event_id = mne.events_from_annotations(raw)

    # Epoching (trial segmentation, e.g., 0–1s after event)
    epochs = mne.Epochs(raw, events, event_id=event_id, tmin=0, tmax=1, baseline=None, preload=True)

    # Extract trial IDs (sample index of each event)
    trial_ids = epochs.events[:, 0]  # Sample index of event onset

    for trial_id, epoch, label in zip(trial_ids, epochs.get_data(), epochs.events[:, -1]):
    # epoch shape: (n_channels, n_times)
        channel_features = [trial_id, file_name.split('.')[0]]  # Extract subject ID from filename
        for channel in epoch:
            channel_features.extend([
                np.std(channel),
                np.mean(channel),
                np.max(channel),
                np.min(channel)
            ])
        features.append(channel_features)
        labels.append(label)

# Transform dataset into pandas dataframe 
df = pd.DataFrame(features,columns=feature_names)
df['label'] = labels

# Analyze dataframe shape, columns, data types, and sample rows
print(df.info())
display(df.describe())
display(df.head())
print(f" dataframe shape {df.shape}")
df['subject_id'].value_counts()

#Save dataframe to csv in data folder
df.to_csv('../data/interim/eeg_motor_imagery.csv',index=False)


# Key Takeaways

**Dimension**
- 30 rows per subject X 259 columns
- Features are all float

**Column Naming Convention**
- Features follow the pattern: `ch{channel_number}_{statistic}`
- Example: `ch1_std`, `ch1_mean`, `ch1_max`, `ch1_min`

**Key Identifiers**
- `subject_id`: File name that uniquely identifies the trial subject
- `trial_id`: Timestamp indicating when the trial occurred

**Note:**
Dataset is shorter than wide, more trials should be added. Update: more files have been added to balance the dataset
