==> notebooks/03_preprocessing.ipynb <==
# Overview

# Import necessary packages
from utils import pd, train_test_split

# Load data from csv file
df = pd.read_csv('../data/interim/eeg_motor_imagery.csv')
df.head()

# Remove rest state data to keep only left and right hand imagery
df = df[df['label'] != 1]
df['label'].value_counts()


# Map labels left hand from 2 → 0, right hand from 3 → 1
df['label'] = df['label'].map({2: 0, 3: 1})
df['label'].value_counts()


# Split data into train and test 
train_df, test_df = train_test_split(df, random_state=42, test_size=0.2)
train_df.shape, test_df.shape

# Save data to processed folder
train_df.to_csv('../data/processed/train_data.csv', index=False)
test_df.to_csv('../data/processed/test_data.csv', index=False)

# Key Takeways

- Rows with a label of 1 indicating rest state have been removed
- Label has been mapped from 2 and 3 to 1 and 0 
- Data has been split into train and test data and saved in the data/processed folder
- No encoding or null handling is needed
