==> notebooks/05_auto_ml_training.ipynb <==
# Overview
- Train an Autogluon predictor (autoML)


# Import necessary packages
from utils import pd, np, TabularPredictor, accuracy_score

# Import data 
train_df = pd.read_csv("../data/processed/train_data.csv")
test_df = pd.read_csv("../data/processed/test_data.csv")

# Set random seed for reproducibility
np.random.seed(42)

# Initialize autogluon tabular predictor
predictor = TabularPredictor(
    label='label',
    eval_metric='accuracy',
    path='../models/autogluon'
)

# Train autogluon tabular predictor
predictor.fit(
    train_data=train_df,
    time_limit=600,  # 10 minutes time limit
    presets='medium_quality_faster_train',  # Use medium quality preset for faster training
    num_bag_folds=5,  # Number of folds for bagging
    num_stack_levels=2,  # Number of stacking levels
    verbosity=2,  # Show detailed training progress
)

# Evaluate model performance
performance = predictor.evaluate(test_df)
performance

# Display autogluon leadorboard
predictor.leaderboard(test_df)

# Display feature importance
feature_importance = predictor.feature_importance(train_df) 
feature_importance.sort_values(by='importance', ascending=False).head()
# runs too long

# Key Takeways

- autogluon performs with an accuracy of 0.76 with a stacked ensemble
